+++ mi_companion/qgis_utilities/__init__.py
@@ -5,3 +5,6 @@ from .expressions import *
 from .paths import *
 from .qgis_logging import *
 from .string_parsing import *
+
+# from .split_detection import *
+from .anchor_centering import *